+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Device alive `list` method to fetch list of devices with alive time
+- Device alive `get` method to fetch one device with alive time
+
 ## [1.10.1] - 2022-05-06
 ### Fixed
 - Incorrect URL path on updating device application version